+++ FlareSwift/FlareSwift/animation/keyframes/keyframe_trigger.swift
@@ -11,10 +11,10 @@ import Foundation
 class KeyFrameTrigger: KeyFrame {
     var _time: Double = 0.0
     
-    func applyInterpolation(component: ActorComponent, time: Double, toFrame: KeyFrame, mix: Double) {
+    func applyInterpolation(component: ActorComponent, time: Double, toFrame: KeyFrame, mix: Float) {
     }
     
-    func apply(component: ActorComponent, mix: Double) {
+    func apply(component: ActorComponent, mix: Float) {
     }
     
     func setNext(_ frame: KeyFrame) {
